Abstract combat simulation, make SparCraft combat simulator using new abstraction interface
