+++ UAlbertaBot/Source/CombatSimulation.cpp
@@ -6,174 +6,36 @@ using namespace UAlbertaBot;
 CombatSimulation::CombatSimulation(
 	shared_ptr<AKBot::OpponentView> opponentView,
 	std::shared_ptr<AKBot::Logger> logger,
-	const BotSparCraftConfiguration& sparcraftConfiguration)
-	: _opponentView(opponentView)
-	, _logger(logger)
-	, _sparcraftConfiguration(sparcraftConfiguration)
+	const BotSparCraftConfiguration& sparcraftConfiguration,
+	const BotMicroConfiguration& microConfiguration)
+	: _sparcraftCombatEstimator(opponentView, logger, sparcraftConfiguration, microConfiguration)
 {
 }
 
 // sets the starting states based on the combat units within a radius of a given position
 // this center will most likely be the position of the forwardmost combat unit we control
-void CombatSimulation::setCombatUnits(
+bool CombatSimulation::isWinPredicted(
 	const std::vector<BWAPI::Unit> ourCombatUnits,
 	std::vector<UnitInfo> enemyCombatUnits,
-	const BWAPI::Position & center,
-	const int radius,
 	int currentFrame)
 {
-	SparCraft::GameState s;
-
-	BWAPI::Broodwar->drawCircleMap(center.x, center.y, 10, BWAPI::Colors::Red, true);
-
-	for (auto & unit : ourCombatUnits)
-	{
-        if (unit->getType().isWorker() || unit->getHitPoints() == 0 || unit->getType().isBuilding())
-        {
-            continue;
-        }
-
-        if (UnitUtil::IsCombatUnit(unit) && SparCraft::System::UnitTypeSupported(unit->getType()))
-		{
-            try
-            {
-			    s.addUnit(getSparCraftUnit(unit, currentFrame));
-            }
-            catch (int e)
-            {
-                e=1;
-                _logger->log("Problem Adding Self Unit with ID: %d", unit->getID());
-            }
-		}
-	}
-
-	for (UnitInfo & ui : enemyCombatUnits)
-	{ 
-        if (ui.type.isWorker() || ui.lastHealth == 0 || ui.type == BWAPI::UnitTypes::Unknown)
-        {
-            continue;
-        }
-        
-        // if it's a bunker that has a nontrivial amount of hit points, replace it by 5 marines
-        if (ui.type == BWAPI::UnitTypes::Terran_Bunker && ui.lastHealth > 10)
-        {
-            double hpRatio = static_cast<double>(ui.lastHealth) / ui.type.maxHitPoints();
-
-            SparCraft::Unit marine( BWAPI::UnitTypes::Terran_Marine,
-                            SparCraft::Position(ui.lastPosition), 
-                            ui.unitID, 
-                            getSparCraftPlayerID(ui.player), 
-                            static_cast<SparCraft::HealthType>(BWAPI::UnitTypes::Terran_Marine.maxHitPoints() * hpRatio),
-                            static_cast<SparCraft::HealthType>(0),
-		                    static_cast<SparCraft::TimeType>(currentFrame), 
-                            static_cast<SparCraft::TimeType>(currentFrame));	
-
-            for (size_t i(0); i < 5; ++i)
-            {
-                s.addUnit(marine);
-            }
-            
-            continue;
-        }
-
-        if (!ui.type.isFlyer() && SparCraft::System::UnitTypeSupported(ui.type) && ui.completed)
-		{
-            try
-            {
-			    s.addUnit(getSparCraftUnit(ui, currentFrame));
-            }
-            catch (int e)
-            {
-                _logger->log("Problem Adding Enemy Unit with ID: %d %d", ui.unitID, e);
-            }
-		}
-	}
-
-	_state = s;
+	return _sparcraftCombatEstimator.isWinPredicted(ourCombatUnits, enemyCombatUnits, currentFrame);
 }
 
-// Gets a SparCraft unit from a BWAPI::Unit, used for our own units since we have all their info
-const SparCraft::Unit CombatSimulation::getSparCraftUnit(BWAPI::Unit unit, int currentFrame) const
+void CombatSimulation::printDebugInformation(BWAPI::Position simulationCenter)
 {
-    return SparCraft::Unit( unit->getType(),
-                            SparCraft::Position(unit->getPosition()), 
-                            unit->getID(), 
-                            getSparCraftPlayerID(unit->getPlayer()), 
-							static_cast<SparCraft::HealthType>(unit->getHitPoints() + unit->getShields()),
-                            static_cast<SparCraft::HealthType>(0),
-		                    static_cast<SparCraft::TimeType>(currentFrame), 
-                            static_cast<SparCraft::TimeType>(currentFrame));	
-}
+	BWAPI::Broodwar->drawCircleMap(simulationCenter.x, simulationCenter.y, 10, BWAPI::Colors::Red, true);
+	std::stringstream ss1;
+	ss1 << "Initial State:\n";
+	ss1 << SparCraft::AITools::StateToStringCompact(_sparcraftCombatEstimator.getSparCraftState()) << "\n\n";
 
-// Gets a SparCraft unit from a UnitInfo struct, needed to get units of enemy behind FoW
-const SparCraft::Unit CombatSimulation::getSparCraftUnit(const UnitInfo & ui, int currentFrame) const
-{
-	BWAPI::UnitType type = ui.type;
+	std::stringstream ss2;
 
-    // this is a hack, treat medics as a marine for now
-	if (type == BWAPI::UnitTypes::Terran_Medic)
-	{
-		type = BWAPI::UnitTypes::Terran_Marine;
-	}
+	ss2 << "Predicted Outcome: " << _sparcraftCombatEstimator.getLastScore() << "\n";
+	ss2 << SparCraft::AITools::StateToStringCompact(_sparcraftCombatEstimator.getEvaluatedState()) << "\n";
 
-    return SparCraft::Unit( ui.type, 
-                            SparCraft::Position(ui.lastPosition), 
-                            ui.unitID, 
-                            getSparCraftPlayerID(ui.player), 
-                            static_cast<SparCraft::HealthType>(ui.lastHealth), 
-                            static_cast<SparCraft::TimeType>(0),
-		                    static_cast<SparCraft::TimeType>(currentFrame), 
-		static_cast<SparCraft::TimeType>(currentFrame));
-}
+	BWAPI::Broodwar->drawTextScreen(150, 200, "%s", ss1.str().c_str());
+	BWAPI::Broodwar->drawTextScreen(300, 200, "%s", ss2.str().c_str());
 
-double CombatSimulation::simulateCombat()
-{
-    try
-    {
-	    SparCraft::GameState originalState(_state);
-        size_t selfID = getSparCraftPlayerID(_opponentView->self());
-        size_t enemyID = getSparCraftPlayerID(_opponentView->defaultEnemy());
-
-        SparCraft::PlayerPtr selfNOK(new SparCraft::Player_AttackClosest(selfID));
-        SparCraft::PlayerPtr enemyNOK(new SparCraft::Player_AttackClosest(enemyID));
-
-		auto& aiParameters = SparCraft::AIParameters::Instance();
-        SparCraft::PlayerPtr p1 =  aiParameters.getPlayer(selfID, _sparcraftConfiguration.CombatSimPlayerName);
-        SparCraft::PlayerPtr p2 =  aiParameters.getPlayer(enemyID, _sparcraftConfiguration.CombatSimPlayerName);
-
-	    SparCraft::Game game (originalState, p1, p2, 2000);
-
-	    game.play();
-	
-		_evaluatedState = game.getState();
-	    _lastScore = SparCraft::Eval::Eval(_evaluatedState, SparCraft::Players::Player_One, SparCraft::EvaluationMethods::LTD2).val();
-        //std::cout << "LTD2: " << SparCraft::Eval::LTD2(g.getState(), 0) << ", " << SparCraft::Eval::LTD2(g.getState(), 1) << "\n";
-        
-	    return _lastScore;
-    }
-    catch (int e)
-    {
-        _logger->log("SparCraft FatalError, simulateCombat() threw");
-
-        return e;
-    }
-}
-
-const SparCraft::GameState & CombatSimulation::getSparCraftState() const
-{
-	return _state;
-}
-
-const size_t CombatSimulation::getSparCraftPlayerID(BWAPI::Player player) const
-{
-	if (player == _opponentView->self())
-	{
-		return SparCraft::Players::Player_One;
-	}
-	else if (player == _opponentView->defaultEnemy())
-	{
-		return SparCraft::Players::Player_Two;
-	}
-
-	return SparCraft::Players::Player_None;
+	BWAPI::Broodwar->drawTextScreen(240, 280, "Combat Sim : %lf", _sparcraftCombatEstimator.getLastScore());
 }
