+++ UAlbertaBot/VisualStudio/AKBot.vcxproj
@@ -67,6 +67,7 @@
     <ClCompile Include="..\Source\ExplorerManager.cpp" />
     <ClCompile Include="..\Source\GameCommander.cpp" />
     <ClCompile Include="..\Source\ScreenCanvas.cpp" />
+    <ClCompile Include="..\Source\simulation\SparCraftCombatEstimator.cpp" />
     <ClCompile Include="..\Source\strategies\BaseStrategyExecutor.cpp" />
     <ClCompile Include="..\Source\strategies\protoss\DarkTemplarRush.cpp" />
     <ClCompile Include="..\Source\strategies\protoss\DragoonRush.cpp" />
@@ -167,6 +168,8 @@
     <ClInclude Include="..\Source\PlayerLocationProvider.h" />
     <ClInclude Include="..\Source\Rect.h" />
     <ClInclude Include="..\Source\ScreenCanvas.h" />
+    <ClInclude Include="..\Source\simulation\CombatEstimator.h" />
+    <ClInclude Include="..\Source\simulation\SparCraftCombatEstimator.h" />
     <ClInclude Include="..\Source\strategies\BaseStrategyExecutor.h" />
     <ClInclude Include="..\Source\strategies\protoss\DarkTemplarRush.h" />
     <ClInclude Include="..\Source\strategies\protoss\DragoonRush.h" />
